+++ windows/lnmp-npm-search.ps1
@@ -1,0 +1,1 @@
+npm search --registry=https://registry.npmjs.org "$args"