+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * `arm32v7` Image based Alpine
 * Switch to PHP 7.3.x
 * Move arm `Dockerfile-arch` to `Dockerfile` folder
+* Update custom script `lnmp-custom-script[.ps1]`
 
 ### Updates: